fix: 응답 수정
- 총 가격 결제 수정
- 응답 매핑 수정

+++ loopz-order-service/src/main/java/kr/co/loopz/order/dto/response/ObjectResponse.java
@@ -4,7 +4,7 @@ public record ObjectResponse(
         String objectId,
         String objectName,
         String imageUrl,
-        int purchasePrice,
+        Long purchasePrice,
         int quantity,
         int stock
 ) {
